requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
numpy>=1.22.0,<1.24.0
ffmpeg-python>=0.2.0
transformers>=4.30.0
sentencepiece>=0.1.99
deep-translator>=1.11.4

